Crossplatform work

+++ NetTarget/Util/FastMemManip.h
@@ -22,6 +22,8 @@
 
 #include "MR_Types.h"
 
+#include <cstdint>
+
 inline void MR_Mem32Set( MR_UInt32* pDest, MR_UInt32 pValue, int pCount )
 {
    for( int lCounter = 0; lCounter < pCount; lCounter++ )
@@ -43,12 +45,12 @@ inline void MR_LargeMem16Set( MR_UInt16* pDest, MR_UInt16 pValue, int pCount )
    if( pCount > 0 )
    {
       MR_UInt32  lValue = pValue|( pValue<<16);
-      MR_UInt32* lDest  = (MR_UInt32*)(((int)pDest)&~1);
+      MR_UInt32* lDest  = (MR_UInt32*)((uintptr_t)pDest & ~(uintptr_t)1);
 
 
       pDest[ pCount-1 ] = pValue;
 
-      if( (int)pDest&1 )
+      if( (uintptr_t)pDest&1 )
       {
          *pDest = pValue; 
          pCount--;
